+++ trinity-util/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2026-05-15
+
+### Added
+- Shared `hotkey` module built on `global-hotkey` for application-wide system shortcut registration.
+- Hotkey config loading and targeted persistence helpers.
+
 ## 2025-01-30
 
 ### Added